Checks for libraries using pkg-config
The build was somehow broken due the fact that the libraries
path was placed hardcoded on the qmake file. Now it is using
the pkg-config to get the path of the libraries and headers.
Note that the dependence of the airprobe was placed. airprobe
location can be configured on the qmake file.

Signed-off-by: Felipe Zimmerle <felipe@zimmerle.org>